requirements.txt
Pillow==10.0.0
django-pyodbc-azure==2.1.0.0
django-background-tasks==1.2.5